requirements.txt
streamlit>=1.32.0
plaid-python>=14.0.0
gspread>=6.0.0
google-auth>=2.28.0
google-generativeai>=0.5.0
pandas>=2.2.0
psutil>=5.9.8
python-dotenv>=1.0.1
sqlcipher3-binary>=0.5.2
cryptography>=42.0.0
requests>=2.31.0
plotly>=5.20.0
